refactor(ui,app): split files to satisfy file-length limit
Extract the resource-usage bar rendering into explorer_resource_usage.go and
the preview-invalidation helpers into update_navigation_invalidate.go so both
source files stay under the 800-line revive limit. Pure move, no behavior
change.
